correct an error

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float e;
-	printf("Size of an char: %lu byte(s)\n",(unsigned long)sizeof(a));
-	printf("Size of an int: %lu byte(s)\n",(unsigned long)sizeof(b));
-	printf("Size of an long int: %lu byte(s)\n",(unsigned long)sizeof(c));
-	printf("Size of an long long: %lu byte(s)\n",(unsigned long)sizeof(d));
-	printf("Size of an float: %lu byte(s)\n",(unsigned long)sizeof(e));
+	printf("Size of an char: %lu byte(s)\n", sizeof(a));
+	printf("Size of an int: %lu byte(s)\n", sizeof(b));
+	printf("Size of an long int: %lu byte(s)\n", sizeof(c));
+	printf("Size of an long long: %lu byte(s)\n", sizeof(d));
+	printf("Size of an float: %lu byte(s)\n", sizeof(e));
 	return (0);
 }
